+++ src/main/java/com/carrot/carrotmarketclonecoding/board/service/impl/BoardServiceImpl.java
@@ -1,21 +1,20 @@
 package com.carrot.carrotmarketclonecoding.board.service.impl;
 
-import static com.carrot.carrotmarketclonecoding.board.domain.QBoard.board;
-
 import com.carrot.carrotmarketclonecoding.board.domain.Board;
 import com.carrot.carrotmarketclonecoding.board.dto.BoardRequestDto.BoardRegisterRequestDto;
 import com.carrot.carrotmarketclonecoding.board.dto.BoardRequestDto.BoardSearchRequestDto;
 import com.carrot.carrotmarketclonecoding.board.dto.BoardRequestDto.BoardUpdateRequestDto;
 import com.carrot.carrotmarketclonecoding.board.dto.BoardRequestDto.MyBoardSearchRequestDto;
 import com.carrot.carrotmarketclonecoding.board.dto.BoardResponseDto.BoardDetailResponseDto;
-import com.carrot.carrotmarketclonecoding.board.dto.BoardResponseDto.BoardNotificationResponseDto;
 import com.carrot.carrotmarketclonecoding.board.dto.BoardResponseDto.BoardSearchResponseDto;
 import com.carrot.carrotmarketclonecoding.board.repository.BoardLikeRepository;
 import com.carrot.carrotmarketclonecoding.board.repository.BoardPictureRepository;
 import com.carrot.carrotmarketclonecoding.board.repository.BoardRepository;
+import com.carrot.carrotmarketclonecoding.board.service.BoardNotificationService;
 import com.carrot.carrotmarketclonecoding.board.service.BoardService;
 import com.carrot.carrotmarketclonecoding.board.service.SearchKeywordRedisService;
 import com.carrot.carrotmarketclonecoding.board.service.VisitRedisService;
+import com.carrot.carrotmarketclonecoding.board.util.HeaderUtil;
 import com.carrot.carrotmarketclonecoding.category.domain.Category;
 import com.carrot.carrotmarketclonecoding.category.repository.CategoryRepository;
 import com.carrot.carrotmarketclonecoding.common.exception.BoardNotFoundException;
@@ -24,16 +23,9 @@ import com.carrot.carrotmarketclonecoding.common.exception.MemberNotFoundExcepti
 import com.carrot.carrotmarketclonecoding.common.exception.TmpBoardNotFoundException;
 import com.carrot.carrotmarketclonecoding.common.exception.UnauthorizedAccessException;
 import com.carrot.carrotmarketclonecoding.common.response.PageResponseDto;
-import com.carrot.carrotmarketclonecoding.keyword.domain.Keyword;
-import com.carrot.carrotmarketclonecoding.keyword.repository.KeywordRepository;
 import com.carrot.carrotmarketclonecoding.member.domain.Member;
 import com.carrot.carrotmarketclonecoding.member.repository.MemberRepository;
-import com.carrot.carrotmarketclonecoding.notification.domain.enums.NotificationType;
-import com.carrot.carrotmarketclonecoding.notification.service.NotificationService;
 import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -51,14 +43,10 @@ public class BoardServiceImpl implements BoardService {
     private final CategoryRepository categoryRepository;
     private final BoardPictureRepository boardPictureRepository;
     private final BoardLikeRepository boardLikeRepository;
-    private final KeywordRepository keywordRepository;
     private final VisitRedisService visitRedisService;
     private final SearchKeywordRedisService searchKeywordRedisService;
     private final BoardPictureService boardPictureService;
-    private final NotificationService notificationService;
-
-    private static final String HEADER_X_FORWARDED_FOR = "X-Forwarded-For";
-    private static final String HEADER_USER_AGENT = "User-Agent";
+    private final BoardNotificationService boardNotificationService;
 
     @Override
     public Long register(Long authId, BoardRegisterRequestDto registerRequestDto, MultipartFile[] pictures, boolean tmp) {
@@ -68,7 +56,7 @@ public class BoardServiceImpl implements BoardService {
         board.setPriceZeroIfMethodIsShare();
         boardRepository.save(board);
         boardPictureService.uploadPicturesIfExistAndUnderLimit(pictures, board);
-        sendKeywordNotification(registerRequestDto, board);
+        boardNotificationService.sendKeywordNotification(registerRequestDto, board);
         return board.getId();
     }
 
@@ -77,9 +65,9 @@ public class BoardServiceImpl implements BoardService {
         Board board = boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
         int like = boardLikeRepository.countByBoard(board);
 
-        String ip = getClientIp(request);
+        String ip = HeaderUtil.getClientIp(request);
         log.debug("IP: {}", ip);
-        String userAgent = getUserAgent(request);
+        String userAgent = HeaderUtil.getUserAgent(request);
         log.debug("USER-AGENT: {}", userAgent);
         if (visitRedisService.increaseVisit(board.getId().toString(), ip, userAgent)) {
             board.increaseVisit();
@@ -159,38 +147,5 @@ public class BoardServiceImpl implements BoardService {
 
     private boolean isLogin(Long authId) {
         return authId != null && authId > 0L;
-    }
-
-    // TODO Util 클래스 분리
-    private String getClientIp(HttpServletRequest request) {
-        String ip = request.getHeader(HEADER_X_FORWARDED_FOR);
-        if (ip == null || ip.isEmpty()) {
-            ip = request.getRemoteAddr();
-        }
-        return ip;
-    }
-
-    // TODO Util 클래스 분리
-    private String getUserAgent(HttpServletRequest request) {
-        return request.getHeader(HEADER_USER_AGENT);
-    }
-
-    // TODO BoardNotificationService 클래스로 분리?
-    private void sendKeywordNotification(BoardRegisterRequestDto registerRequestDto, Board board) {
-        Set<String> wordsNotDuplicated = getTitleAndDescriptionNotDuplicated(registerRequestDto);
-        Set<Keyword> keywords = keywordRepository.findByNameIn(wordsNotDuplicated);
-        BoardNotificationResponseDto notification = new BoardNotificationResponseDto(board);
-        sendBoardNotificationResponseDto(keywords, notification);
-    }
-
-    private Set<String> getTitleAndDescriptionNotDuplicated(BoardRegisterRequestDto registerRequestDto) {
-        String[] words = (registerRequestDto.getTitle() + " " + registerRequestDto.getDescription()).split(" ");
-        return new HashSet<>(List.of(words));
-    }
-
-    private void sendBoardNotificationResponseDto(Set<Keyword> keywords, BoardNotificationResponseDto notification) {
-        for (Keyword keyword : keywords) {
-            notificationService.add(keyword.getMember().getAuthId(), NotificationType.NOTICE, notification);
-        }
     }
 }
